Fix type-limits warning again

+++ cpp/PlusOne.cpp
@@ -6,11 +6,11 @@ class Solution {
  public:
   vector<int> plusOne(vector<int> &digits) {
     unsigned long n = digits.size() - 1;
-    for (unsigned long i = n; i >= 0; i--) {
-      if (digits[i] == 9) {
-        digits[i] = 0;
+    for (int i = n; i >= 0; i--) {
+      if (digits[(unsigned long)i] == 9) {
+        digits[(unsigned long)i] = 0;
       } else {
-        digits[i] += 1;
+        digits[(unsigned long)i] += 1;
         return digits;
       }
     }
